Add initial values to the database

+++ app/src/main/java/com/pgaa/teamvelocity/data/entity/Sprint.kt
@@ -12,7 +12,7 @@ data class Sprint (
     @PrimaryKey @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "quarter") val quarter: String,
     @ColumnInfo(name = "description") val description: String,
-    @ColumnInfo(name = "story_points") val storyPoints: Int,
-    @ColumnInfo(name = "man_day") val mandDay: Int,
+    @ColumnInfo(name = "story_points") val storyPoints: Double,
+    @ColumnInfo(name = "man_day") val mandDay: Double,
     @ColumnInfo(name = "ratio") val ratio: Double
 ) : Parcelable
